requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
cssselect==1.1.0
cssutils==2.6.0
cycler==0.11.0
fonttools==4.37.1
idna==3.3
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.5.3
numpy==1.23.2
packaging==21.3
Pillow==9.2.0
premailer==3.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
urllib3==1.26.12
yagmail==0.15.285
